+++ src/components/member/__tests__/user-avatar.test.js
@@ -2,13 +2,13 @@ import React from 'react'
 import { shallow } from 'enzyme'
 import Avatar from '@material-ui/core/Avatar'
 
-import { mockEampleUser1, mockEampleUser3 } from '../../../__mocks__/mocks'
+import { mockExampleUser1, mockExampleUser3 } from '../../../__mocks__/mocks'
 
 import UserAvatar from '../user-avatar'
 
 const props = {
   member: {
-    ...mockEampleUser1,
+    ...mockExampleUser1,
     avatarImg: 'http://example.com/avatar.jpg',
   },
 }
@@ -19,7 +19,7 @@ describe('Component/UserAvatar', () => {
   })
 
   test('should render an user w/o an avatar without throwing an error', () => {
-    expect(shallow(<UserAvatar member={mockEampleUser3} />).dive()).toMatchSnapshot()
+    expect(shallow(<UserAvatar member={mockExampleUser3} />).dive()).toMatchSnapshot()
   })
 
   test('should render an Avatar', () => {