+++ src/components/trello-card/index.js
@@ -2,6 +2,7 @@ import { connect } from 'react-redux'
 import { find, get } from 'lodash'
 import Card from './component'
 import { isCardHidden } from './utils'
+import { toggleMinimizeLabels } from '../../actions/app'
 import { addMembersEstimation } from '../../actions/members'
 import { addBoardEstimations } from '../../actions/boards'
 
@@ -9,6 +10,7 @@ const mapStateToProps = state => ({
   boardData: get(state, 'boards.data', []),
   memberToggle: get(state, 'app.memberToggle', {}),
   members: get(state, 'members.members', []),
+  minimizeLabels: get(state, 'app.trelloCardUI.minimizeLabels') || false,
 })
 
 const mapDispatchToProps = dispatch => ({ dispatch })
@@ -34,10 +36,12 @@ const mergeProps = (stateProps, { dispatch }, ownProps) => {
 
   return {
     ...ownProps,
+    addEstimations,
     boardName,
     isHidden: isCardHidden(isCardHiddenProps),
-    addEstimations,
+    minimizeLabels: stateProps.minimizeLabels,
     toggleList: stateProps.toggleList,
+    toggleMinimizeLabels: () => dispatch(toggleMinimizeLabels()),
   }
 }